pyproject.toml
[project]
name = "ebd-planner"
version = "0.1.0"
description = ""
authors = [
    {name = "ErickViniciux",email = "erickmilhomemv@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "prisma (>=0.15.0,<0.16.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "ebd_planner", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
